Use a JSON apir rather than artifacts
